// digits-admin-backend/src/dto/purchaseManagement.dto.ts
export class PurchaseManagementDTO {
  plate_id: string;
  plate_number: string;
  buyer_id: string;
  owner_id: string;
  bid_price: string;
  sell_price: string;
  orderID: string;
  trans_id: string;
  payment_status: string;
  buy_type: string;
  runner_status: number;
  delivery_status: number;
  runner_id: string;

  price: string;
  Total: string;
  transaction_fee: string;
  expires: string;
  name: string;
  email: string;
  phone: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  zipCode: string;
  order: string;
  created_at: Date;
}